chore: move to correct paths
